+++ client/errors.go
@@ -21,9 +21,11 @@ var (
 
 	requestIdRegex = regexp.MustCompile(`\s([Rr]equest[ _]{0,1}(ID|Id|id):)\s[A-Za-z0-9-]+`)
 	hostIdRegex    = regexp.MustCompile(`\sHostID: [A-Za-z0-9+/_=-]+`)
-	arnIdRegex     = regexp.MustCompile(`(\s)(arn:aws[A-Za-z0-9-]*:)[^ \.\(\)\[\]\{\}\;\,]+(\s?)`)
-	urlRegex       = regexp.MustCompile(`([\s"])http(s?):\/\/[a-z0-9_\-\./]+([":\s]?)`)
-	lookupRegex    = regexp.MustCompile(
+	arnIdRegex     = regexp.MustCompile(
+		`\b(arn:aws[A-Za-z0-9-]*:)` + // "arn:aws:"
+			`[^ \.\(\)\[\]\{\}\;\,'"]+`) // continuation in the form of a bunch of characters except space, dot, brackets, square brackets, curly brackets, semicolon, comma and single/double quotes
+	urlRegex    = regexp.MustCompile(`([\s"])http(s?):\/\/[a-z0-9_\-\./]+([":\s]?)`)
+	lookupRegex = regexp.MustCompile(
 		`\blookup\s[-A-Za-z0-9\.]+\s` + // " lookup host.name "
 			`on\s\S+:\d+`, // "on 123.123.123.123:53"
 	)
@@ -35,11 +37,13 @@ var (
 		`\bdial\s(tcp|udp)\s` + // "dial tcp "
 			fmt.Sprintf(`(?:%s|%s):\d+`, ipv4Regex, bracketedIpv6Regex), // "192.168.1.2:123" or "[::1]:123"
 	)
-	encAuthRegex           = regexp.MustCompile(`(\s)(Encoded authorization failure message:)\s[A-Za-z0-9_-]+`)
-	userRegex              = regexp.MustCompile(`(\s)(is not authorized to perform: .+ on resource:\s)(user)\s.+`)
-	s3Regex                = regexp.MustCompile(`(\s)(S3(Key|Bucket))=(.+?)([,;\s])`)
-	resourceNotExistsRegex = regexp.MustCompile(`(\sThe )([A-Za-z0-9 -]+ )'([A-Za-z0-9-]+?)'( does not exist)`)
-	resourceNotFoundRegex  = regexp.MustCompile(`([A-Za-z0-9 -]+)( name not found - Could not find )([A-Za-z0-9 -]+)( named )'([A-Za-z0-9-]+?)'`)
+	ec2ImageIdRegex          = regexp.MustCompile(`The image ID '[^']+'`)
+	encAuthRegex             = regexp.MustCompile(`(\s)(Encoded authorization failure message:)\s[A-Za-z0-9_-]+`)
+	userRegex                = regexp.MustCompile(`(\s)(is not authorized to perform: .+ on resource:\s)(user)\s.+`)
+	s3Regex                  = regexp.MustCompile(`(\s)(S3(Key|Bucket))=(.+?)([,;\s])`)
+	resourceNotExistsRegex   = regexp.MustCompile(`(\sThe )([A-Za-z0-9 -]+ )'([A-Za-z0-9-]+?)'( does not exist)`)
+	resourceNotFoundRegex    = regexp.MustCompile(`([A-Za-z0-9 -]+)( name not found - Could not find )([A-Za-z0-9 -]+)( named )'([A-Za-z0-9-]+?)'`)
+	autoscalingGroupNotFound = regexp.MustCompile(`(ValidationError: Group ).+( not found)`)
 )
 
 var errorCodeDescriptions = map[string]string{
@@ -51,6 +55,7 @@ var errorCodeDescriptions = map[string]string{
 	"AccessDenied":                  "You are not authorized to perform this operation. Check your IAM policies, and ensure that you are using the correct access keys.",
 	"AuthorizationError":            "You are not authorized to perform this operation. Check your IAM policies, and ensure that you are using the correct access keys.",
 	"AuthFailure":                   "You are not authorized to perform this operation. Check your IAM policies, and ensure that you are using the correct access keys.",
+	"InvalidToken":                  "The provided token is malformed or otherwise invalid.",
 }
 
 var throttleCodes = map[string]struct{}{
@@ -74,7 +79,7 @@ func classifyError(err error, fallbackType diag.Type, accounts []string, opts ..
 	var ae smithy.APIError
 	if errors.As(err, &ae) {
 		switch ae.ErrorCode() {
-		case "AccessDenied", "AccessDeniedException", "UnauthorizedOperation", "AuthorizationError", "OptInRequired", "SubscriptionRequiredException", "InvalidClientTokenId", "AuthFailure", "ExpiredToken", "FailedResourceAccessException":
+		case "AccessDenied", "AccessDeniedException", "UnauthorizedOperation", "AuthorizationError", "OptInRequired", "SubscriptionRequiredException", "InvalidClientTokenId", "AuthFailure", "ExpiredToken", "ExpiredTokenException", "FailedResourceAccessException", "InvalidToken":
 			return diag.Diagnostics{
 				RedactError(accounts, diag.NewBaseError(err,
 					diag.ACCESS,
@@ -260,7 +265,7 @@ func removePII(aa []string, msg string) string {
 	}
 	msg = requestIdRegex.ReplaceAllString(msg, " ${1} xxxx")
 	msg = hostIdRegex.ReplaceAllString(msg, " HostID: xxxx")
-	msg = arnIdRegex.ReplaceAllString(msg, "${1}${2}xxxx${3}")
+	msg = arnIdRegex.ReplaceAllString(msg, "${1}xxxx${2}")
 	msg = urlRegex.ReplaceAllString(msg, "${1}http${2}://xxxx${3}")
 	msg = lookupRegex.ReplaceAllString(msg, "lookup xxxx on xxxx:xx")
 	msg = readXonYRegex.ReplaceAllString(msg, "read $1 xxxx:xx->xxxx:xx")
@@ -270,6 +275,8 @@ func removePII(aa []string, msg string) string {
 	msg = s3Regex.ReplaceAllString(msg, "${1}${2}=xxxx${5}")
 	msg = resourceNotExistsRegex.ReplaceAllString(msg, "${1}${2}'xxxx'${4}")
 	msg = resourceNotFoundRegex.ReplaceAllString(msg, "${1}${2}${3}${4}'xxxx'")
+	msg = ec2ImageIdRegex.ReplaceAllString(msg, "The image ID 'xxxx'")
+	msg = autoscalingGroupNotFound.ReplaceAllString(msg, "${1}xxxx${2}")
 	msg = accountObfusactor(aa, msg)
 
 	return msg