+++ tools/tool.go
@@ -5,4 +5,6 @@ package main
 
 import (
 	_ "github.com/cloudquery/cq-gen"
+	_ "github.com/golang/mock/mockgen"
+	_ "github.com/golang/mock/mockgen/model"
 )